Extracting pip installs and data generation into separate notebooks
- New "notebook-config" notebook with pip installs and helper functions
- New "generate-data" notebook with data generation code
- Extracting boilerplate code into helper functions
- Formating "skeleton" notebook
